Rename image_dir_fixture.py -> dataset_fixture.py
