+++ tests/invoke_training/_shared/data/data_loaders/test_dreambooth_sd_dataloader.py
@@ -6,7 +6,7 @@ from invoke_training._shared.data.data_loaders.dreambooth_sd_dataloader import (
 from invoke_training.config.data.data_loader_config import AspectRatioBucketConfig, DreamboothSDDataLoaderConfig
 from invoke_training.config.data.dataset_config import ImageDirDatasetConfig
 
-from ..image_dir_fixture import image_dir  # noqa: F401
+from ..dataset_fixtures import image_dir  # noqa: F401
 
 
 def test_build_dreambooth_sd_dataloader(image_dir):  # noqa: F811
