+++ tests/invoke_training/_shared/data/datasets/test_image_pair_preference_dataset.py
@@ -2,7 +2,7 @@ import PIL.Image
 
 from invoke_training._shared.data.datasets.image_pair_preference_dataset import ImagePairPreferenceDataset
 
-from ..image_dir_fixture import image_pair_preference_dir  # noqa: F401
+from ..dataset_fixtures import image_pair_preference_dir  # noqa: F401
 
 
 def test_image_dir_dataset_len(image_pair_preference_dir):  # noqa: F811
